This was an attempt at soft collision by calculating multiple vectors. I realised it wasn't worth it to add to the project since it wasn't optimised right and I didn't have enough time to optimise it right. So because of this I decided to make another branch specific for this which I may continue in my personal time in the near future

+++ src/core/Vector2D.java
@@ -9,7 +9,7 @@ public class Vector2D {
         this.y = y;
     }
 
-    public void multiply(int speed) {
+    public void multiply(double speed) {
         x *= speed;
         y *= speed;
     }
